requirements.txt
python-docx
spacy
pymupdf
openai
regex
unoconv
reportlab
extract-msg
flair
torch
docx2pdf
streamlit
